requirements.txt
#Python version 3.12.4
matplotlib==3.9.1
numpy==1.26.4
pillow==10.4.0
tbb==2021.13.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
torch==2.3.1+cu121
torchvision==0.18.1
tqdm==4.66.4

